Define {u,}int32_t to long
For some reason original Xtensa support patch contains a hack to define {u,}int32_t to be int rather than long.
This makes no difference in practice as both are 32-bit but it breaks formatting constants,
e.g. PRIu32 is still defined to long. This causes -Wformat warnings.
We revert this bit.